requirements.txt
numpy>=1.21.0
torch>=2.0.0
matplotlib>=3.7.0
pandas>=1.5.0
scipy>=1.10.0
scikit-learn>=1.2.0
xgboost>=1.7.0
pyyaml>=6.0
tqdm>=4.64.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.0.0
